Add hook manage modal state
