+++ pages/core-unit/[code]/finances/reports/index.tsx
@@ -1,3 +1,4 @@
+import { fetchExpenseCategories } from '@ses/containers/FinancesOverview/api/queries';
 import request from 'graphql-request';
 import React, { useState, useEffect } from 'react';
 import { GRAPHQL_ENDPOINT } from '../../../../../src/config/endpoints';
@@ -6,14 +7,16 @@ import { fetchCoreUnits } from '../../../../../src/stories/components/CoreUnitSu
 import { TransparencyReport } from '../../../../../src/stories/containers/TransparencyReport/TransparencyReport';
 import { CORE_UNIT_REQUEST } from '../../../../../src/stories/containers/TransparencyReport/transparencyReportAPI';
 import type { CoreUnitDto } from '../../../../../src/core/models/dto/coreUnitDTO';
+import type { ExpenseCategory } from '@ses/core/models/dto/expenseCategoriesDTO';
 import type { GetServerSidePropsContext } from 'next';
 
 interface TransparencyProps {
   coreUnits: CoreUnitDto[];
   cu: CoreUnitDto;
+  expenseCategories: ExpenseCategory[];
 }
 
-const Transparency = ({ coreUnits, cu }: TransparencyProps) => {
+const Transparency = ({ coreUnits, cu, expenseCategories }: TransparencyProps) => {
   const [currentCoreUnit, setCurrentCoreUnit] = useState<CoreUnitDto>(cu);
   useEffect(() => {
     setCurrentCoreUnit(cu);
@@ -27,7 +30,7 @@ const Transparency = ({ coreUnits, cu }: TransparencyProps) => {
         coreUnits,
       }}
     >
-      <TransparencyReport coreUnits={coreUnits} coreUnit={currentCoreUnit} />
+      <TransparencyReport coreUnits={coreUnits} coreUnit={currentCoreUnit} expenseCategories={expenseCategories} />
     </CoreUnitContext.Provider>
   );
 };
@@ -39,7 +42,11 @@ export const getServerSideProps = async (context: GetServerSidePropsContext) => 
   const code = query.code as string;
   const { query: gqlQuery, filter } = CORE_UNIT_REQUEST(code);
 
-  const [data, coreUnits] = await Promise.all([request(GRAPHQL_ENDPOINT, gqlQuery, filter), fetchCoreUnits()]);
+  const [data, coreUnits, expenseCategories] = await Promise.all([
+    request(GRAPHQL_ENDPOINT, gqlQuery, filter),
+    fetchCoreUnits(),
+    fetchExpenseCategories(),
+  ]);
 
   if (data?.coreUnits?.length === 0 || code === 'DEL') {
     return {
@@ -51,6 +58,7 @@ export const getServerSideProps = async (context: GetServerSidePropsContext) => 
     props: {
       coreUnits,
       cu: data.coreUnits[0],
+      expenseCategories,
     },
   };
 };
