+++ src/stories/containers/TransparencyReport/components/TransparencyActuals/useTransparencyActuals.ts
@@ -1,3 +1,4 @@
+import { useModalCategory } from '@ses/components/BasicModal/useModalCategory';
 import { API_MONTH_TO_FORMAT } from '@ses/core/utils/date';
 import { capitalizeSentence, getWalletWidthForWallets, toKebabCase } from '@ses/core/utils/string';
 import _ from 'lodash';
@@ -15,11 +16,13 @@ import {
 import { ACTUALS_BREAKDOWN_QUERY_PARAM } from '../../utils/constants';
 import type { InnerTableColumn, InnerTableRow } from '@ses/components/AdvancedInnerTable/AdvancedInnerTable';
 import type { BudgetStatementDto, BudgetStatementWalletDto } from '@ses/core/models/dto/coreUnitDTO';
+import type { ExpenseCategory } from '@ses/core/models/dto/expenseCategoriesDTO';
 import type { DateTime } from 'luxon';
 
 export const useTransparencyActuals = (
   propsCurrentMonth: DateTime,
-  budgetStatements: BudgetStatementDto[] | undefined
+  budgetStatements: BudgetStatementDto[] | undefined,
+  expenseCategories: ExpenseCategory[]
 ) => {
   const currentMonth = useMemo(() => propsCurrentMonth.toFormat(API_MONTH_TO_FORMAT), [propsCurrentMonth]);
   const router = useRouter();
@@ -253,14 +256,24 @@ export const useTransparencyActuals = (
     wallets,
   ]);
 
+  const {
+    openModal,
+    handleChangeItemAccordion,
+    handleCheckedExpandedAll,
+    handleCloseModal,
+    handleOpenModal,
+    headCountCategory,
+    notHeadCountCategory,
+    checkOut,
+  } = useModalCategory(expenseCategories);
   const [breakdownColumnsForActiveTab, allBreakdownColumns] = useMemo(() => {
     const allBreakdownColumns: { [key: string]: InnerTableColumn[] } = {};
     for (const wallet of wallets) {
-      allBreakdownColumns[wallet?.name] = getActualsBreakdownColumns(wallet);
+      allBreakdownColumns[wallet?.name] = getActualsBreakdownColumns(wallet, handleOpenModal);
     }
 
     return [allBreakdownColumns[currentWallet?.name], allBreakdownColumns];
-  }, [currentWallet, wallets]);
+  }, [currentWallet?.name, handleOpenModal, wallets]);
 
   const allBreakdownItems = useMemo(() => {
     const result: { [key: string]: InnerTableRow[] } = {};
@@ -288,5 +301,12 @@ export const useTransparencyActuals = (
     mainTableItems,
     breakdownTabs,
     wallets,
+    openModal,
+    handleCheckedExpandedAll,
+    handleCloseModal,
+    headCountCategory,
+    notHeadCountCategory,
+    handleChangeItemAccordion,
+    checkOut,
   };
 };
